This is sender side code, nothing is different from receiver but to be sure

+++ src/call/rtp_transport_controller_send.h
@@ -144,7 +144,7 @@ class RtpTransportControllerSend final
                                           int64_t now_ms,int pathid)
       RTC_RUN_ON(task_queue_);
   void PostUpdates(NetworkControlUpdate update,NetworkControlUpdate update2) RTC_RUN_ON(task_queue_);
-  int MpFindBestPath(int64_t rtt1,int64_t rtt2,double loss1,double loss2,int64_t rate1,int64_t rate2) RTC_RUN_ON(task_queue_);
+  int MpFindBestPath(int64_t rtt1,int64_t rtt2,double loss1,double loss2,double rate1,double rate2) RTC_RUN_ON(task_queue_);
   void MpFindRatio(int64_t rtt1,int64_t rtt2,int64_t rate1,int64_t rate2,int pathid) RTC_RUN_ON(task_queue_);
   void UpdateControlState() RTC_RUN_ON(task_queue_);
   RtpPacketPacer* pacer();
